Move context to top level datalab.Context, remove Project and Projects
