+++ google/datalab/__init__.py
@@ -10,3 +10,4 @@
 # or implied. See the License for the specific language governing permissions and limitations under
 # the License.
 
+from _context import Context
