+++ tests/bigquery/api_tests.py
@@ -16,8 +16,8 @@ import unittest
 import mock
 from oauth2client.client import AccessTokenCredentials
 
+import google.datalab
 import google.datalab.bigquery
-import google.datalab.context
 import google.datalab.utils
 from google.datalab.bigquery._api import Api
 
@@ -393,4 +393,4 @@ class TestCases(unittest.TestCase):
   def _create_context():
     project_id = 'test'
     creds = AccessTokenCredentials('test_token', 'test_ua')
-    return google.datalab.context.Context(project_id, creds)
+    return google.datalab.Context(project_id, creds)
