+++ tests/bigquery/table_tests.py
@@ -21,8 +21,8 @@ from oauth2client.client import AccessTokenCredentials
 import pandas
 import unittest
 
+import google.datalab
 import google.datalab.bigquery
-import google.datalab.context
 import google.datalab.utils
 
 
@@ -673,7 +673,7 @@ class TestCases(unittest.TestCase):
   def _create_context():
     project_id = 'test'
     creds = AccessTokenCredentials('test_token', 'test_ua')
-    return google.datalab.context.Context(project_id, creds)
+    return google.datalab.Context(project_id, creds)
 
   @staticmethod
   def _create_table(name):
