+++ tests/bigquery/udf_tests.py
@@ -15,8 +15,8 @@ from __future__ import unicode_literals
 from oauth2client.client import AccessTokenCredentials
 import unittest
 
+import google.datalab
 import google.datalab.bigquery
-import google.datalab.context
 
 
 class TestCases(unittest.TestCase):
@@ -58,4 +58,4 @@ class TestCases(unittest.TestCase):
   def _create_context():
     project_id = 'test'
     creds = AccessTokenCredentials('test_token', 'test_ua')
-    return google.datalab.context.Context(project_id, creds)
+    return google.datalab.Context(project_id, creds)
