+++ tests/stackdriver/monitoring/metric_tests.py
@@ -17,7 +17,7 @@ import unittest
 
 import google.cloud.monitoring
 
-import google.datalab.context
+import google.datalab
 import google.datalab.stackdriver.monitoring as gcm
 
 PROJECT = 'my-project'
@@ -41,7 +41,7 @@ class TestCases(unittest.TestCase):
     self.context = self._create_context()
     self.descriptors = gcm.MetricDescriptors(context=self.context)
 
-  @mock.patch('google.datalab.context._context.Context.default')
+  @mock.patch('google.datalab.Context.default')
   def test_constructor_minimal(self, mock_context_default):
     mock_context_default.return_value = self.context
 
@@ -161,7 +161,7 @@ class TestCases(unittest.TestCase):
   @staticmethod
   def _create_context(project_id='test'):
     creds = AccessTokenCredentials('test_token', 'test_ua')
-    return google.datalab.context.Context(project_id, creds)
+    return google.datalab.Context(project_id, creds)
 
   @staticmethod
   def _list_metrics_get_result(context):
