+++ tests/storage/object_tests.py
@@ -16,7 +16,7 @@ import mock
 from oauth2client.client import AccessTokenCredentials
 import unittest
 
-import google.datalab.context
+import google.datalab
 import google.datalab.storage
 import google.datalab.utils
 
@@ -87,7 +87,7 @@ class TestCases(unittest.TestCase):
   def _create_context():
     project_id = 'test'
     creds = AccessTokenCredentials('test_token', 'test_ua')
-    return google.datalab.context.Context(project_id, creds)
+    return google.datalab.Context(project_id, creds)
 
   @staticmethod
   def _create_objects_get_result():
